refactor: 참가자가 자신의 카드를 기반으로 버스트 여부를 검사하는 기능 추가
