+++ src/test/kotlin/blackjack/model/DealerTest.kt
@@ -2,6 +2,8 @@ package blackjack.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.ValueSource
 
 class DealerTest {
 
@@ -32,5 +34,18 @@ class DealerTest {
         dealer.addCard(Card(CardNumber.TWO, Suit.HEART))
         val actual = dealer.judge(participant)
         assertThat(actual).isEqualTo(GameResult.WIN)
+    }
+
+
+    @ParameterizedTest
+    @ValueSource(ints = [14, 15, 16])
+    fun `카드 총 합이 16 이하면 true를 반환한다`(sum: Int) {
+        val actual = isHitable(sum)
+        assertThat(actual).isEqualTo(true)
+    }
+
+    fun isHitable(score: Int): Boolean {
+        val threshold = 17
+        return score < threshold
     }
 }
